requirements.txt
arrow==0.12.1
astroid==1.6.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.2
django-cors-headers==2.2.0
django-filter==1.1.0
django-import-export==1.0.0
django-rest-swagger==2.2.0
django-routers==0.2
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
et-xmlfile==1.0.1
Flask==0.12.2
gunicorn==19.8.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.12
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.2
psycopg2==2.7.4
PyJWT==1.5.3
pylint==1.8.3
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
simplejson==3.13.2
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.2.0
Werkzeug==0.14.1
whitenoise==3.3.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
